requirements.txt
Django==6.0.3
django-filter==25.2
python-decouple==3.8
psycopg2-binary==2.9.11
gunicorn==25.1.0
whitenoise==6.12.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
PyJWT==2.11.0
django-cors-headers==4.9.0
dj-rest-auth==7.1.1
redis==7.3.0
channels==4.3.2
channels-redis==4.3.0
daphne==4.2.1
gevent==25.9.1
asgiref==3.11.1
Pillow==12.1.1
opencv-python==4.13.0.92
opencv-contrib-python==4.13.0.92
pytest==9.0.2
pytest-django==4.12.0
pytest-cov==7.0.0
pytest-asyncio==1.3.0
pytest-xdist==3.8.0
celery==5.6.2
django-simple-history==3.11.0
numpy==2.4.3
django-axes==8.3.1
